OE0-10: Views stored as submodule, not single python file
