+++ api_fhir_r4/serializers/__init__.py
@@ -13,8 +13,8 @@ class BaseFHIRSerializer(serializers.Serializer):
     fhirConverter = BaseFHIRConverter()
 
     def __init__(self, *args, **kwargs):
+        self._reference_type = kwargs.pop('reference_type', ReferenceConverterMixin.UUID_REFERENCE_TYPE)
         super().__init__(*args, **kwargs)
-        self._reference_type = kwargs.get('reference_type', ReferenceConverterMixin.UUID_REFERENCE_TYPE)
 
     def to_representation(self, obj):
         if isinstance(obj, HttpResponseBase):
