+++ api_fhir_r4/urls.py
@@ -3,40 +3,48 @@ from django.urls import include, path
 from rest_framework.routers import DefaultRouter
 from openIMIS.openimisapps import *
 
+from api_fhir_r4.views import LoginView, fhir as fhir_viewsets
 
 imis_modules = openimis_apps()
 
 
 router = DefaultRouter()
-router.register(r'login', views.LoginView, basename="login")
-# register endpojnt related to Location module if used
+router.register(r'login', LoginView, basename="login")
+# register endpoint related to Location module if used
 if 'location' in imis_modules:
-    router.register(r'Location', views.LocationViewSet, basename="Location_R4")
-    router.register(r'HealthcareService', views.HealthcareServiceViewSet, basename="HealthcareService_R4")
+    router.register(r'Location', fhir_viewsets.LocationViewSet, basename="Location_R4")
+    router.register(r'HealthcareService', fhir_viewsets.HealthcareServiceViewSet, basename="HealthcareService_R4")
+
 # register endpoint for insuree if used
 if 'insuree' in imis_modules:
-    router.register(r'Patient', views.InsureeViewSet, basename="Patient_R4")
-    router.register(r'Group', views.GroupViewSet, basename="Group_R4")
+    router.register(r'Patient', fhir_viewsets.InsureeViewSet, basename="Patient_R4")
+    router.register(r'Group', fhir_viewsets.GroupViewSet, basename="Group_R4")
     router.register(
-        r'CoverageEligibilityRequest', views.CoverageEligibilityRequestViewSet, basename="CoverageEligibilityRequest_R4"
+        r'CoverageEligibilityRequest', fhir_viewsets.CoverageEligibilityRequestViewSet, basename="CoverageEligibilityRequest_R4"
     )
+
 # register endpoints related to medical module
 if 'medical' in imis_modules:
-    router.register(r'Medication', views.MedicationViewSet, basename="Medication_R4")
-    router.register(r'Condition', views.ConditionViewSet, basename="Condition_R4")
-    router.register(r'ActivityDefinition', views.ActivityDefinitionViewSet, basename="ActivityDefinition_R4")
-# register all endpoints related to claim based on claim
+    router.register(r'Medication', fhir_viewsets.MedicationViewSet, basename="Medication_R4")
+    router.register(r'Condition', fhir_viewsets.ConditionViewSet, basename="Condition_R4")
+    router.register(r'ActivityDefinition', fhir_viewsets.ActivityDefinitionViewSet, basename="ActivityDefinition_R4")
+
+# register all endpoints related to c based on c
 if 'claim' in imis_modules:
-    router.register(r'Claim', views.ClaimViewSet, basename="Claim_R4")
-    router.register(r'ClaimResponse', views.ClaimResponseViewSet, basename="ClaimResponse_R4")
-    router.register(r'PractitionerRole', views.PractitionerRoleViewSet, basename="PractitionerRole_R4")
-    router.register(r'Practitioner', views.PractitionerViewSet, basename="Practitioner_R4")
-    router.register(r'CommunicationRequest', views.CommunicationRequestViewSet, basename="CommunicationRequest_R4")
+    router.register(r'Claim', fhir_viewsets.ClaimViewSet, basename="Claim_R4")
+    router.register(r'ClaimResponse', fhir_viewsets.ClaimResponseViewSet, basename="ClaimResponse_R4")
+    router.register(r'PractitionerRole', fhir_viewsets.PractitionerRoleViewSet, basename="PractitionerRole_R4")
+    router.register(r'Practitioner', fhir_viewsets.PractitionerViewSet, basename="Practitioner_R4")
+    router.register(r'CommunicationRequest', fhir_viewsets.CommunicationRequestViewSet, basename="CommunicationRequest_R4")
+
 # register endpoint for policy if used
 if 'policy' in imis_modules:
-    router.register(r'Coverage', views.CoverageRequestQuerySet, basename="Coverage_R4")
-    router.register(r'Contract', views.ContractViewSet, basename="Contract_R4")
+    router.register(r'Coverage', fhir_viewsets.CoverageRequestQuerySet, basename="Coverage_R4")
+    router.register(r'Contract', fhir_viewsets.ContractViewSet, basename="Contract_R4")
+
 # register endpoint for policy holder if used
 if 'policyholder' in imis_modules:
-    router.register(r'Organisation', views.OrganisationViewSet, basename="Organisation_R4")
+    router.register(r'Organisation', fhir_viewsets.OrganisationViewSet, basename="Organisation_R4")
+
+
 urlpatterns = [path('', include(router.urls))]
