+++ api_fhir_r4/views/fhir/__init__.py
@@ -1,0 +1,16 @@
+from api_fhir_r4.views.fhir.activity_definition import ActivityDefinitionViewSet
+from api_fhir_r4.views.fhir.claim import ClaimViewSet
+from api_fhir_r4.views.fhir.claim_response import ClaimResponseViewSet
+from api_fhir_r4.views.fhir.communication_request import CommunicationRequestViewSet
+from api_fhir_r4.views.fhir.condition import ConditionViewSet
+from api_fhir_r4.views.fhir.contract import ContractViewSet
+from api_fhir_r4.views.fhir.coverage_eligibility_request import CoverageEligibilityRequestViewSet
+from api_fhir_r4.views.fhir.coverage_request import CoverageRequestQuerySet
+from api_fhir_r4.views.fhir.group import GroupViewSet
+from api_fhir_r4.views.fhir.health_service import HealthcareServiceViewSet
+from api_fhir_r4.views.fhir.insuree import InsureeViewSet
+from api_fhir_r4.views.fhir.location import LocationViewSet
+from api_fhir_r4.views.fhir.medication import MedicationViewSet
+from api_fhir_r4.views.fhir.organisation import OrganisationViewSet
+from api_fhir_r4.views.fhir.practitioner import PractitionerViewSet
+from api_fhir_r4.views.fhir.practitioner_role import PractitionerRoleViewSet
