+++ api_fhir_r4/views/fhir/coverage_request.py
@@ -1,0 +1,47 @@
+import datetime
+
+from policy.models import Policy
+
+from rest_framework import mixins
+from rest_framework.viewsets import GenericViewSet
+
+from api_fhir_r4.permissions import FHIRApiCoverageRequestPermissions
+from api_fhir_r4.serializers.coverageSerializer import CoverageSerializer
+from api_fhir_r4.views.fhir.fhir_base_viewset import BaseFHIRView
+
+
+class CoverageRequestQuerySet(BaseFHIRView, mixins.RetrieveModelMixin, mixins.ListModelMixin,mixins.UpdateModelMixin,mixins.CreateModelMixin,GenericViewSet):
+    lookup_field = 'uuid'
+    serializer_class = CoverageSerializer
+    permission_classes = (FHIRApiCoverageRequestPermissions,)
+
+    def list(self, request, *args, **kwargs):
+        queryset = self.get_queryset()
+        queryset.prefetch_related('services')
+        refDate = request.GET.get('refDate')
+        refEndDate = request.GET.get('refEndDate')
+        identifier = request.GET.get("identifier")
+        if identifier:
+            queryset = queryset.filter(chf_id=identifier)
+        else:
+            queryset = queryset.filter(validity_to__isnull=True).order_by('validity_from')
+            if refDate != None:
+                isValidDate = True
+                try :
+                    datevar = datetime.datetime.strptime(refDate, "%Y-%m-%d").date()
+                except ValueError :
+                    isValidDate = False
+                queryset = queryset.filter(validity_from__gte=datevar)
+            if refEndDate != None:
+                isValidDate = True
+                try :
+                    datevar = datetime.datetime.strptime(refEndDate, "%Y-%m-%d").date()
+                except ValueError :
+                    isValidDate = False
+                queryset = queryset.filter(validity_from__lt=datevar)
+
+        serializer = CoverageSerializer(self.paginate_queryset(queryset), many=True)
+        return self.get_paginated_response(serializer.data)
+
+    def get_queryset(self):
+        return Policy.get_queryset(None, self.request.user)
